# Optimize point increment by using projections