```
goos: linux
goarch: amd64
pkg: github.com/AlexanderYastrebov/wireguard-vanity-key
cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
                      │    main     │                HEAD                 │
                      │   sec/op    │   sec/op     vs base                │
FindBatchPoint/1024-8   374.6n ± 2%   231.8n ± 0%  -38.13% (p=0.000 n=10)

                      │    main    │              HEAD              │
                      │    B/op    │    B/op     vs base            │
FindBatchPoint/1024-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                      │    main    │              HEAD              │
                      │ allocs/op  │ allocs/op   vs base            │
FindBatchPoint/1024-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```