@AlexanderYastrebov 2025-02-08
I think this might be close to the optimum meaning future performance gains would be minimal.

![Screenshot from 2025-02-08 01-06-00](https://github.com/user-attachments/assets/e677dd0a-a36a-4b85-b505-866f5ce892e1)